+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to Semantic Versioning(http://semver.org/).
 
 ## 3.0.1
-- Fix the issue that currency format was incorrect
+- Fixed the issue that currency format was incorrect
+- Fixed the issue that configuration key was too long on prestashop 1.5 
 
 ## 3.0.0
 ### Added