+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to Semantic Versioning(http://semver.org/).
 
 ## 3.0.1
 - Fixed the issue that currency format was incorrect
-- Fixed the issue that configuration key was too long on prestashop 1.5 
+- Fixed the issue that configuration key was too long on prestashop 1.5
+- Fixed the product variation id is missing when multi-currency is enabled
 
 ## 3.0.0
 ### Added